requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
emoji==2.8.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
pyparsing==3.1.1
python-dotenv==0.21.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
uritemplate==4.1.1
urllib3==1.26.6
